+++ lib/taurus/cli/test/__init__.py
@@ -21,25 +21,4 @@
 ##
 #############################################################################
 
-"""
-This module provides the taurus Command Line Interface
-
-It is based on the click module to provide CLI utilities.
-
-It defines a `taurus` command which can accept subcommands defined in
-other taurus submodules or even in 3rd party plugins.
-
-To define a subcommand for taurus, you need to register it in the
-`taurus.cli.subcommands` entry point group via setuptools.
-
-.. todo:: add click link , add code snippet for plugins, etc.
-
-This module provides also common options/flags for command-line interfaces
-that can be used by taurus plugins. For list of available options,
-see :mod:`taurus.cli.common`
-
-.. seealso:: :mod:`taurus.cli.alt`
-
-"""
-
-from .cli import main, register_subcommands, taurus_cmd
+"""Tests for taurus.cli"""